Add virtio-blk driver with block device layer
- Add BlockDevice trait for sector-based I/O in resource/block.rs
- Add BlockDeviceWrapper for byte-to-sector translation with alignment handling
- Create virtio_block.rs driver using existing VirtioHal
- Add BlockScheme for block:/pci/XX:XX.X device paths
- Change SEEK_SET/CUR/END constants from usize to u32
- Add kernel integration tests (8 tests)
- Add userspace integration test (block_test)
- Update TODO.md with block device discovery as future work

+++ panda-abi/src/lib.rs
@@ -134,9 +134,9 @@ pub const OP_SURFACE_UPDATE_PARAMS: u32 = 0x6_0004;
 // =============================================================================
 
 // Seek whence values
-pub const SEEK_SET: usize = 0;
-pub const SEEK_CUR: usize = 1;
-pub const SEEK_END: usize = 2;
+pub const SEEK_SET: u32 = 0;
+pub const SEEK_CUR: u32 = 1;
+pub const SEEK_END: u32 = 2;
 
 // =============================================================================
 // Shared types
